Initial commit.
This project will be for users to submit files for display at church's via their presentation software.
Either;
Pre Service videos
Sermon Slides
Announcement, videos and images

Requiring specific formats, and resolutions. Or converting the files if possible.
MacOS requires powerpoint files to be converted to seperate images to be displayed and controled by ProPresentor.
Videos to be widescreen, and mp4 preferably. and all images to also be 16x9, at either 1920x1080 or 1280x720.

With the additional configuration to be able to have cutoff times, for submissions, ie thursday night cutoff for the coming
sunday, (if submitted after this, it will be placed into the list for the following week.)

Included, also the possibility of creating the 'presentor' files to drop stright into the presentor program (not requiring
importing), with the addition of a possibly desktop or cli app, to 'sync' the server and the computer the service.